Extract generateToken into a utils function
Also changes payload for a successful login to `{userId: Number, isAdmin: Boolean}`. The client isn't going to be able to decrypt the payload unless they know the secret — and they shouldn't know the secret. The payload is for the server, for subsequent requests. And all the server needs on subsequent requests is the `userId`.

Signed-off-by: Miikis <28997062+Miikis@users.noreply.github.com>